Include more variance in odd eggs
Low chance that the egg has a pokemon that is not shiny, low chance that the pokemon has all perfect DVs, but high chance that it has one perfect DV and is therefore shiny. Chance that it has no perfect DVs but also has all around better DVs than nonperfect shinies.