Remove surf check for wild encounters
The game adds variety to surfing levels by randomly increasing the level
of the encounter by 1-4, allowing a 5 level variance, despite having
only 3 hardcoded possibilities per route. By removing this check, we can
now remove the hardcoded grass encounter levels.

The side effect of this change is that for low-level grass encounters we
will need to have the "level" set to 2. This should be fine as it should
be easy enough to run from wild encounters with a starter at level 5 and
a wild pidgey at level 6, but will keep an eye on this.